# Switch to msgpackr as faster alternative