--- a/client/lib/package.json
+++ b/client/lib/package.json
@@ -50,11 +50,11 @@
     "dependencies": {
         "amazon-chime-sdk-js": "^3.18.2",
         "buffer": "^6.0.3",
+        "component-emitter": "^2.0.0",
         "localforage": "^1.10.0",
-        "notepack.io": "~2.2.0",
+        "msgpackr": "^1.11.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "socket.io-client": "^4.7.2",
-        "socket.io-msgpack-parser": "^3.0.2"
+        "socket.io-client": "^4.7.2"
     }
 }