… client/demo/dist/index.js