[msgpackr](https://www.npmjs.com/package/msgpackr) offers superior performance across all amounts of data among all JS implementations of MessagePack. Benchmarks are taken from [notepack.io](https://github.com/darrachequesne/notepack/tree/main/benchmarks) and nodejs version is v21.7.1

encode.js benchmark
```
+----------------------------+-------------------+-------------------+----------------+---------------+
|                            │ tiny              │ small             │ medium         │ large         |
+----------------------------+-------------------+-------------------+----------------+---------------+
| notepack                   │ 4,265,278 ops/sec │ 908,909 ops/sec   │ 67,599 ops/sec │ 662 ops/sec   |
+----------------------------+-------------------+-------------------+----------------+---------------+
| msgpack-js                 │ 246,378 ops/sec   │ 200,327 ops/sec   │ 20,406 ops/sec │ 270 ops/sec   |
+----------------------------+-------------------+-------------------+----------------+---------------+
| msgpack-lite               │ 730,040 ops/sec   │ 319,070 ops/sec   │ 40,140 ops/sec │ 712 ops/sec   |
+----------------------------+-------------------+-------------------+----------------+---------------+
| @msgpack/msgpack           │ 977,182 ops/sec   │ 562,043 ops/sec   │ 56,765 ops/sec │ 207 ops/sec   |
+----------------------------+-------------------+-------------------+----------------+---------------+
| msgpackr                   │ 6,091,010 ops/sec │ 1,260,195 ops/sec │ 57,931 ops/sec │ 829 ops/sec   |
+----------------------------+-------------------+-------------------+----------------+---------------+
| JSON.stringify (to Buffer) │ 1,963,163 ops/sec │ 424,998 ops/sec   │ 25,987 ops/sec │ 40.95 ops/sec |
+----------------------------+-------------------+-------------------+----------------+---------------+
```

decode.js benchmark
```
+--------------------------+-------------------+-------------------+----------------+-------------+
|                          │ tiny              │ small             │ medium         │ large       |
+--------------------------+-------------------+-------------------+----------------+-------------+
| notepack                 │ 4,721,660 ops/sec │ 1,033,792 ops/sec │ 52,844 ops/sec │ 407 ops/sec |
+--------------------------+-------------------+-------------------+----------------+-------------+
| msgpack-js               │ 1,560,533 ops/sec │ 351,532 ops/sec   │ 30,389 ops/sec │ 363 ops/sec |
+--------------------------+-------------------+-------------------+----------------+-------------+
| msgpack-lite             │ 963,938 ops/sec   │ 210,107 ops/sec   │ 19,407 ops/sec │ 290 ops/sec |
+--------------------------+-------------------+-------------------+----------------+-------------+
| @msgpack/msgpack         │ 2,217,199 ops/sec │ 804,214 ops/sec   │ 48,447 ops/sec │ 337 ops/sec |
+--------------------------+-------------------+-------------------+----------------+-------------+
| msgparck                 │ 9,227,912 ops/sec │ 1,686,493 ops/sec │ 74,785 ops/sec │ 411 ops/sec |
+--------------------------+-------------------+-------------------+----------------+-------------+
| JSON.parse (from Buffer) │ 1,914,721 ops/sec │ 548,322 ops/sec   │ 43,427 ops/sec │ 100 ops/sec |
+--------------------------+-------------------+-------------------+----------------+-------------+
```